examples/avatar_agents/audio_wave/requirements.txt
fastapi
uvicorn
opencv-python
livekit-agents[google,elevenlabs]~=1.0
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.0rc
librosa
python-dotenv
